+++ backend/ddang/src/main/java/com/ddang/ddang/authentication/configuration/Oauth2WebConfiguration.java
@@ -1,0 +1,34 @@
+package com.ddang.ddang.authentication.configuration;
+
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.format.FormatterRegistry;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+@RequiredArgsConstructor
+public class Oauth2WebConfiguration implements WebMvcConfigurer {
+
+    private final Oauth2TypeConverter oauth2TypeConverter;
+    private final AuthenticationInterceptor authenticationInterceptor;
+    private final AuthenticationPrincipalArgumentResolver authenticationArgumentResolver;
+
+    @Override
+    public void addFormatters(final FormatterRegistry registry) {
+        registry.addConverter(oauth2TypeConverter);
+    }
+
+    @Override
+    public void addInterceptors(final InterceptorRegistry registry) {
+        registry.addInterceptor(authenticationInterceptor)
+                .addPathPatterns("/**");
+    }
+
+    @Override
+    public void addArgumentResolvers(final List<HandlerMethodArgumentResolver> resolvers) {
+        resolvers.add(authenticationArgumentResolver);
+    }
+}
